cloud_function/requirements.txt
flask==3.0.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.18.1
google-generativeai==0.3.1
pandas==2.1.4
openpyxl==3.1.2
dbt-bigquery==1.7.2
gunicorn==21.2.0 